+++ shared-resources-mc20/src/main/java/nl/enjarai/shared_resources/mc20/impl/ScreenElementsBuilderImpl.java
@@ -1,9 +1,14 @@
 package nl.enjarai.shared_resources.mc20.impl;
 
+import me.shedaniel.clothconfig2.api.AbstractConfigListEntry;
 import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.text.Text;
+import nl.enjarai.shared_resources.mc20.gui.DirectoryConfigEntry;
 import nl.enjarai.shared_resources.mc20.gui.LongButtonWidget;
 import nl.enjarai.shared_resources.versioned.ScreenElementsBuilder;
+
+import java.nio.file.Path;
+import java.util.function.Consumer;
 
 public class ScreenElementsBuilderImpl implements ScreenElementsBuilder {
     @Override
@@ -16,5 +21,10 @@ public class ScreenElementsBuilderImpl implements ScreenElementsBuilder {
     @Override
     public ButtonWidget buildLongButton(int x, int y, int width, int height, Text text, ButtonWidget.PressAction onPress) {
         return new LongButtonWidget(x, y, width, height, text, onPress);
+    }
+
+    @Override
+    public AbstractConfigListEntry<Path> getDirectoryEntry(Text fieldName, Path initialValue, Path defaultValue, Consumer<Path> saveConsumer) {
+        return new DirectoryConfigEntry(fieldName, initialValue, defaultValue, saveConsumer);
     }
 }